+++ lib/super/badge.rb
@@ -12,43 +12,19 @@ module Super
       purple: "bg-purple-800 text-white",
     }
 
-    def initialize(text, style: nil, styles: nil)
+    def initialize(text, style:)
+      if !STYLES.key?(style)
+        raise Super::Error::Initialization, "`style:` must be one of #{STYLES.keys.inspect}"
+      end
+
       @text = text
-      if styles.present?
-        Super::Useful::Deprecation["0.22"].deprecation_warning("styles:", "use `style:` with a single style")
-      end
-      @requested_styles = Array(styles.presence).flatten + Array(style.presence).flatten
-      if @requested_styles.any? { |s| s.is_a?(String) }
-        Super::Useful::Deprecation["0.22"].warn("Super::Badge.new(text, style:) accepts exactly one Symbol style from this list #{STYLES.keys.inspect}")
-      end
-      if @requested_styles.size != 1
-        Super::Useful::Deprecation["0.22"].warn("Super::Badge.new(text, style:) accepts exactly one style, but it received #{@requested_styles.size}")
-      end
+      @style = style
     end
 
     def styles
       return @styles if instance_variable_defined?(:@styles)
-      @styles =
-        if requested_styles.delete(:reset)
-          []
-        else
-          COMMON_STYLES
-        end
 
-      if requested_styles.empty?
-        @styles += [STYLES[:light]]
-      else
-        requested_styles.each do |style|
-          @styles +=
-            if STYLES.key?(style)
-              [STYLES[style]]
-            else
-              [style]
-            end
-        end
-      end
-
-      @styles
+      @styles = COMMON_STYLES + [STYLES.fetch(@style)]
     end
 
     def to_s