+++ test/lib/super/badge_test.rb
@@ -5,18 +5,13 @@ class BadgeTest < ActiveSupport::TestCase
     badge = build("text", style: :light)
 
     assert_includes(badge.attr("class"), Super::Badge::STYLES[:light])
+    assert_includes(badge.attr("class"), Super::Badge.const_get(:COMMON_STYLES).join(" "))
   end
 
-  test %(Badge.new("text", styles: :light)) do
-    badge = build("text", styles: :light)
-
-    assert_includes(badge.attr("class"), Super::Badge::STYLES[:light])
-  end
-
-  test %(Badge.new("text", styles: ["custom custom-1"])) do
-    badge = build("text", styles: ["custom", "custom-1"])
-
-    assert_includes(badge.attr("class"), "custom", "custom-1")
+  test %(Badge.new("test", style: :doesnt_exist)) do
+    assert_raises(Super::Error::Initialization) do
+      build("text", style: :doesnt_exist)
+    end
   end
 
   private